+++ boardgamegeek/collection.py
@@ -1,3 +1,15 @@
+# coding: utf-8
+"""
+:mod:`boardgamegeek.collection` - Collection information
+========================================================
+
+.. module:: boardgamegeek.collection
+   :platform: Unix, Windows
+   :synopsis: classes for storing collection information
+
+.. moduleauthor:: Cosmin Luță <q4break@gmail.com>
+
+"""
 from __future__ import unicode_literals
 from copy import copy
 
@@ -7,7 +19,12 @@ from .utils import DictObject
 
 
 class Collection(DictObject):
+    """
+    A dictionary-like object represeting a ``Collection``
 
+    :param dict data: a dictionary containing the collection data
+    :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekError` in case of invalid data
+    """
     def __init__(self, data):
         kw = copy(data)
 
@@ -38,6 +55,12 @@ class Collection(DictObject):
             log.info("")
 
     def add_game(self, game):
+        """
+        Add a game to the ``Collection``
+
+        :param dict game: game data
+        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekError` in case of invalid data
+        """
         try:
             # Collections can have duplicate elements (different collection ids), so don't add the same thing multiple times
             if game["id"] not in self.__game_ids:
@@ -61,10 +84,22 @@ class Collection(DictObject):
 
     @property
     def owner(self):
+        """
+        Return the collection's owner
+
+        :returns: the collection's owner
+        :rtype: str
+        """
         return self._data.get("owner")
 
     @property
     def items(self):
+        """
+        Returns the items in the collection
+
+        :returns: the items in the collection
+        :rtype: list of :py:class:`boardgamegeek.games.CollectionBoardGame`
+        """
         return self._items
 
     def __iter__(self):